refactor: update user authentication flow by integrating user profile retrieval and removing deprecated API routes
